Trait podio::Endianness

pub trait Endianness {
    fn int_to_target<T: EndianConvert>(val: T) -> T;

    fn int_from_target<T: EndianConvert>(val: T) -> T;
}

Trait implementing conversion methods for a specific endianness

Required methods

fn int_to_target<T: EndianConvert>(val: T) -> T

Converts a value from the platform type to the specified endianness

fn int_from_target<T: EndianConvert>(val: T) -> T

Converts a value from the sepcified endianness to the platform type

Loading content...